fix: fix

+++ src/service/allocation/AllocationServiceFactory.ts
@@ -8,7 +8,9 @@ import { IGetAllocationsService } from '../../domain/service/allocation/getAlloc
 import { ISaveAllocationService } from '../../domain/service/allocation/saveAllocation/ISaveAllocationService'
 import { IDeleteAllocationService } from '../../domain/service/allocation/deleteAllocation/IDeleteAllocationService'
 import { IUpdateAllocationService } from '../../domain/service/allocation/updateAllocation/IUpdateAllocationService'
+import { makeAllocationValidation } from './AllocationValidation'
 
+let allocationService: IAllocationService = null
 export interface IAllocationService {
   getAllocationService: IGetAllocationService
   getAllocationsService: IGetAllocationsService
@@ -18,17 +20,22 @@ export interface IAllocationService {
 }
 
 export const makeAllocationService = (): IAllocationService => {
-  const getAllocationService = makeGetAllocationService().getAllocationService
-  const saveAllocationService = makeSaveAllocationService().saveAllocationService
-  const updateAllocationService = makeUpdateAllocationService().updateAllocationService
-  const deleteAllocationService = makeDeleteAllocationService().deleteAllocationService
-  const getAllocationsService = makeGetAllocationsService().getAllocationsService
+  if (!allocationService) {
+    const validator = makeAllocationValidation()
+    const getAllocationService = makeGetAllocationService().getAllocationService
+    const saveAllocationService = makeSaveAllocationService(validator).saveAllocationService
+    const updateAllocationService = makeUpdateAllocationService(validator).updateAllocationService
+    const deleteAllocationService = makeDeleteAllocationService().deleteAllocationService
+    const getAllocationsService = makeGetAllocationsService().getAllocationsService
 
-  return {
-    getAllocationService,
-    saveAllocationService,
-    updateAllocationService,
-    deleteAllocationService,
-    getAllocationsService
+    allocationService = {
+      getAllocationService,
+      saveAllocationService,
+      updateAllocationService,
+      deleteAllocationService,
+      getAllocationsService
+    }
   }
+
+  return allocationService
 }
